neo_python.py - factor out Utils
